@@ -10,7 +10,7 @@
         }
     ],
     "require": {
-        "php": ">=5.3.0"
+        "php": ">=7.1"
     },
     "require-dev": {
         "phpunit/phpunit": "~6"
